refactor(deployments): move interfaces out of deployments.service.ts
Interfaces used to define the shape of Deployments backend API responses are moved to
deployment-api.service.ts so that they are next to the functions that return them. Two unused
interfaces (Environment and Pods) are removed entirely. NetworkStat is moved to models to fit with
CpuStat and MemoryStat.

related to https://github.com/openshiftio/openshift.io/issues/2962
