+++ src/app/space/create/deployments/models/cpu-stat.ts
@@ -1,3 +1,3 @@
 import { Stat } from './stat';
 
-export declare interface CpuStat extends Stat {}
+export interface CpuStat extends Stat {}
