+++ src/app/space/create/deployments/models/memory-stat.ts
@@ -2,6 +2,6 @@ import { Stat } from './stat';
 
 export type MemoryUnit = 'bytes' | 'KB' | 'MB' | 'GB';
 
-export declare interface MemoryStat extends Stat {
+export interface MemoryStat extends Stat {
   readonly units: MemoryUnit;
 }
