+++ src/app/space/create/deployments/models/network-stat.ts
@@ -1,0 +1,9 @@
+export interface NetworkStat {
+  readonly sent: NetStat;
+  readonly received: NetStat;
+}
+
+export interface NetStat {
+  readonly used: number;
+  readonly timestamp?: number;
+}
