+++ src/app/space/create/deployments/models/pods.ts
@@ -2,7 +2,7 @@ import { PodPhase } from './pod-phase';
 
 export type PodsData = [PodPhase, number];
 
-export declare interface Pods {
+export interface Pods {
   readonly pods: PodsData[];
   readonly total: number;
 }
