+++ src/app/space/create/deployments/models/scaled-net-stat.spec.ts
@@ -1,16 +1,16 @@
-import { ScaledNetworkStat } from './scaled-network-stat';
+import { ScaledNetStat } from './scaled-net-stat';
 
-describe('ScaledNetworkStat', () => {
+describe('ScaledNetStat', () => {
 
   it('should not scale 500 bytes', () => {
-    let stat = new ScaledNetworkStat(500);
+    let stat = new ScaledNetStat(500);
     expect(stat.raw).toEqual(500);
     expect(stat.used).toEqual(500);
     expect(stat.units).toEqual('bytes');
   });
 
   it('should scale 2048 bytes', () => {
-    let stat = new ScaledNetworkStat(2048);
+    let stat = new ScaledNetStat(2048);
     expect(stat.raw).toEqual(2048);
     expect(stat.used).toEqual(2);
     expect(stat.units).toEqual('KB');
@@ -18,7 +18,7 @@ describe('ScaledNetworkStat', () => {
 
   it('should scale 5.5GB', () => {
     let gb = Math.pow(1024, 3);
-    let stat = new ScaledNetworkStat(5.5 * gb);
+    let stat = new ScaledNetStat(5.5 * gb);
     expect(stat.raw).toEqual(5.5 * gb);
     expect(stat.used).toEqual(5.5);
     expect(stat.units).toEqual('GB');
