+++ src/app/space/create/deployments/models/stat.ts
@@ -1,4 +1,4 @@
-export declare interface Stat {
+export interface Stat {
   readonly used: number;
   readonly quota: number;
   readonly timestamp?: number;
